requirements.txt
fastapi[standard]
uvicorn
openai
google-generativeai
pydantic